+++ src/app/app-routing.module.ts
@@ -4,12 +4,14 @@ import { DartPresetComponent } from './dart-preset/dart-preset.component';
 import { DartGameSingleOutComponent } from './dart-game-single-out/dart-game-single-out.component';
 import { DartGameDoubleOutComponent } from './dart-game-double-out/dart-game-double-out.component';
 import { DartGameAroundTheClockComponent } from './dart-game-around-the-clock/dart-game-around-the-clock.component';
+import { DartGameLuckyNumbersComponent } from './dart-game-lucky-numbers/dart-game-lucky-numbers.component';
 
 const routes: Routes = [
   { path: '', component: DartPresetComponent },
   { path: 'singleOut', component: DartGameSingleOutComponent },
   { path: 'doubleOut', component: DartGameDoubleOutComponent },
   { path: 'aroundTheClock', component: DartGameAroundTheClockComponent },
+  { path: 'luckyNumber', component: DartGameLuckyNumbersComponent},
 ];
 
 @NgModule({