+++ src/app/dart-game-lucky-numbers/dart-game-lucky-numbers.component.spec.ts
@@ -1,0 +1,23 @@
+import { ComponentFixture, TestBed } from '@angular/core/testing';
+
+import { DartGameLuckyNumbersComponent } from './dart-game-lucky-numbers.component';
+
+describe('DartGameLuckyNumbersComponent', () => {
+  let component: DartGameLuckyNumbersComponent;
+  let fixture: ComponentFixture<DartGameLuckyNumbersComponent>;
+
+  beforeEach(async () => {
+    await TestBed.configureTestingModule({
+      declarations: [DartGameLuckyNumbersComponent]
+    })
+    .compileComponents();
+    
+    fixture = TestBed.createComponent(DartGameLuckyNumbersComponent);
+    component = fixture.componentInstance;
+    fixture.detectChanges();
+  });
+
+  it('should create', () => {
+    expect(component).toBeTruthy();
+  });
+});