+++ src/app/interfaces/game-data-lucky-number.ts
@@ -1,0 +1,7 @@
+export interface GameDataLuckyNumber {
+    player: string;
+    thrownNumbers: number;
+    trys: number;
+    isActive: boolean;
+    [key: string]: any;
+}