+++ src/assets/changelogs/changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## [1.6.0] - 2025-01-11
+- Hinzugefügt: Neuer Spielmodus "Lucky Number".
 - Änderung: Name der Software wurde von "Dart Scorer" zu "PeiDu Scorer" geändert.
 - Änderung: Das Design für die Spieleinstellungen wurde komplett überarbeitet.
 - Änderung: Im Modus "Around the Clock" können Spieler nun Aktiv/Inaktiv geschaltet werden.